requirements.txt
Pillow==9.4.0
colored==1.4.4
isort==5.12.0